+++ NEWS.md
@@ -1,3 +1,9 @@
+# mrgda development
+
+## New features and changes
+
+- `query_src_list` now searches across values in addition to column and label names. (#251)
+
 # mrgda 0.13.0
 
 ## New features and changes